requirements.txt
tensorflow==2.7.0
streamlit==1.9.2
tensorflow_hub==0.12.0
numpy==1.22.3
matplotlib==3.5.2
Pillow==9.0.1